backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg[binary]>=3.1
pydantic>=2.5
python-dotenv>=1.0
PyJWT>=2.8.0
